# services/api-gateway/docs/openapi.yaml
openapi: 3.0.3
info:
  title: Kalendas API Gateway
  description: |
    Especificación OpenAPI del API Gateway de Kalendas. Todas las operaciones se exponen bajo el gateway y se encaminan a los microservicios internos (calendars, events, comments, notifications).

    Notas:
    - Todas las fechas están en ISO 8601.
    - Los parámetros de búsqueda se pasan como query string en el endpoint raíz de cada recurso.
    - Las respuestas devuelven `id` (string) en lugar de `_id`.
  version: 1.0.0
  contact:
    name: Equipo Kalendas
servers:
  - url: http://localhost:8080
    description: Desarrollo local (API Gateway)
...
paths:
...
  /api/notifications:
    get:
      tags: [Notifications]
      summary: Listar notificaciones con filtros
      description: |
        Filtros admitidos:
        - `eventId`, `calendarId`, `recipientEmail`, `channel`.
        - `organizer`: filtra por eventos cuyo `organizer` coincide.
        - `unread=true`: únicamente no leídas (para `in-app`).
      parameters:
        - in: query
          name: eventId
          schema: { type: string }
        - in: query
          name: calendarId
          schema: { type: string }
        - in: query
          name: recipientEmail
          schema: { type: string, format: email }
        - in: query
          name: organizer
          schema: { type: string }
        - in: query
          name: channel
          schema: { type: string, enum: ['email','in-app'] }
        - in: query
          name: unread
          schema: { type: string, enum: ['true','false'] }
      responses:
        '200':
          description: Lista de notificaciones
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/Notification' }
    post:
      tags: [Notifications]
      summary: Crear notificación de comentario
      description: Genera una notificación asociada a un comentario sobre un evento. El servicio completa mensaje, canal y destinatario.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/NotificationCreate' }
            examples:
              ejemplo:
                value:
                  eventId: 64f1a0c2a0b1c2d3e4f56789
                  calendarId: 64f1a0c2a0b1c2d3e4f56780
                  commenterName: ana.romero
      responses:
        '201': { description: Creado, content: { application/json: { schema: { $ref: '#/components/schemas/Notification' } } } }
  /api/notifications/{id}:
    get:
      tags: [Notifications]
      summary: Obtener notificación por ID
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      responses:
        '200': { description: OK, content: { application/json: { schema: { $ref: '#/components/schemas/Notification' } } } }
        '404': { description: No encontrado }
    delete:
      tags: [Notifications]
      summary: Eliminar notificación por ID
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      responses:
        '204': { description: Eliminado }
        '404': { description: No encontrado }
  /api/notifications/{id}/read:
    patch:
      tags: [Notifications]
      summary: Marcar notificación como leída
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      responses:
        '200': { description: OK, content: { application/json: { schema: { $ref: '#/components/schemas/Notification' } } } }
        '404': { description: No encontrado }
components:
  schemas:
...
    Notification:
      type: object
      properties:
        id: { type: string }
        eventId: { type: string }
        calendarId: { type: string }
        channel: { type: string, enum: [email, in-app] }
        recipientEmail: { type: string, format: email }
        message: { type: string }
        read: { type: boolean }
        status: { type: string, enum: [pending, sent, error] }
        attempts: { type: integer }
        lastError: { type: string, nullable: true }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
    NotificationCreate:
      type: object
      required: [eventId, calendarId, commenterName]
      properties:
        eventId: { type: string }
        calendarId: { type: string }
        commenterName: { type: string }